fix: build for Xcode 16

+++ Sources/SentryCrash/Recording/Monitors/SentryCrashMonitor_System.h
@@ -39,7 +39,7 @@ SentryCrashMonitorAPI *sentrycrashcm_system_getAPI(void);
 
 bool sentrycrash_isSimulatorBuild(void);
 
-uint64_t sentrycrashcm_system_freememory(void);
+unsigned long long sentrycrashcm_system_freememory(void);
 
 #ifdef __cplusplus
 }
